2021.12.09.17.50

+++ src/main/java/com/nova/controllers/UserListCont.java
@@ -1,12 +1,8 @@
 package com.nova.controllers;
 
 import com.nova.models.Games;
-import com.nova.models.enums.Role;
 import com.nova.models.Users;
-import org.springframework.security.authentication.AnonymousAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
+import com.nova.models.enums.Role;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -21,15 +17,7 @@ public class UserListCont extends Main {
 
     @GetMapping("/userList")
     public String userList(Model model) {
-        Users userFromDB = new Users();
-
-        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-        if ((!(auth instanceof AnonymousAuthenticationToken)) && auth != null) {
-            UserDetails userDetail = (UserDetails) auth.getPrincipal();
-            if (userDetail != null) userFromDB = repoUsers.findByUsername(userDetail.getUsername());
-        }
-
-        List<Games> games = repoGames.findAllByUserid(userFromDB.getId());
+        List<Games> games = repoGames.findAllByUserid(checkUser().getId());
         float income = 0;
         for (Games g : games) income += g.getIncome();
 
@@ -43,11 +31,20 @@ public class UserListCont extends Main {
     }
 
     @PostMapping("/userList/{id}/edit")
-    public String userUpdate(
-            @PathVariable(value = "id") Long id,@RequestParam String username,
-            @RequestParam String password, @RequestParam Role role
+    public String userUpdate(Model model,
+                             @PathVariable(value = "id") Long id, @RequestParam String username,
+                             @RequestParam String password, @RequestParam Role role
     ) {
         Users temp = repoUsers.findById(id).orElseThrow();
+        if (temp == checkUser())
+            if (String.valueOf(temp.getRole()).equals(checkUserRole())) {
+                temp.setUsername(username);
+                temp.setPassword(password);
+                temp.setRole(role);
+                repoUsers.save(temp);
+                return "redirect:/index";
+            }
+
         temp.setUsername(username);
         temp.setPassword(password);
         temp.setRole(role);
@@ -56,7 +53,20 @@ public class UserListCont extends Main {
     }
 
     @PostMapping("/userList/{id}/delete")
-    public String userDelete(@PathVariable(value = "id") Long id) {
+    public String userDelete(Model model, @PathVariable(value = "id") Long id) {
+        Users temp = repoUsers.findById(id).orElseThrow();
+        if (temp == checkUser()) {
+            List<Games> games = repoGames.findAllByUserid(checkUser().getId());
+            float income = 0;
+            for (Games g : games) income += g.getIncome();
+            List<Users> users = repoUsers.findAll();
+            model.addAttribute("income", income);
+            model.addAttribute("games", games);
+            model.addAttribute("users", users);
+            model.addAttribute("role", checkUserRole());
+            model.addAttribute("message", "Вы не можете удалить себя!");
+            return "userList";
+        }
         repoUsers.deleteById(id);
         return "redirect:/userList";
     }
